msdk: add the capability to indicate the driver to allocate linear memory
Intel GPU supports tiling and linear memory and the media driver
allocates tiling memory for a surfaces for most cases. Sometimes a tiling
surface will cause some issues when this surface is shared with the 3rd
party library/device. So this patch added the capability to indicate the
driver to allocate linear memory