MSDK: Allow to create context from external display.
The MSDK plugins can now create its context based on the external
display from the other VA based plugins. The pipeline such as:

  gst-launch-1.0 -vf videotestsrc ! video/x-raw,width=800,height=600 \
    ! vaapih264enc ! msdkh264dec ! fakesink

can work with the same VA display for both VA and MSDK plugins.

Before this, VA and MSDK plugins have their own VA displays, which
is verbose. And more important, when there are multi GPU devices and
even one GPU device may have several GPU nodes, it is a little hard
to ensure all of the VA based plugins working on the same device/node
pair. If not, the performance is very low.